split ffi meas/trans into modules
drops extra generic on BaseStability
