+++ rust/opendp/src/meas/stability.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 use std::hash::Hash;
 
-use num::{Integer, Float, NumCast};
+use num::{Integer, Float, NumCast, Zero};
 
 use crate::core::{Measurement, Function, PrivacyRelation, Metric, SensitivityMetric};
 use crate::dist::{L1Sensitivity, L2Sensitivity, SmoothedMaxDivergence};
@@ -11,40 +11,40 @@ use crate::error::Fallible;
 
 // TIK: Type of Input Key
 // TIC: Type of Input Count
-// TOC: Type of Output Count
+// TOC: Type of Output Count (equal to MI::Distance)
 
 pub type CountDomain<TIK, TIC> = SizedDomain<MapDomain<AllDomain<TIK>, AllDomain<TIC>>>;
 
 // tie metric space with distribution
-pub trait BaseStabilityNoise<TOC>: Metric<Distance=TOC> {
-    fn noise(shift: TOC, scale: TOC, enforce_constant_time: bool) -> Fallible<TOC>;
+pub trait BaseStabilityNoise: Metric {
+    fn noise(shift: Self::Distance, scale: Self::Distance, enforce_constant_time: bool) -> Fallible<Self::Distance>;
 }
-impl<TOC: SampleLaplace> BaseStabilityNoise<TOC> for L1Sensitivity<TOC> {
-    fn noise(shift: TOC, scale: TOC, enforce_constant_time: bool) -> Fallible<TOC> {
-        TOC::sample_laplace(shift, scale, enforce_constant_time)
+impl<TOC: SampleLaplace> BaseStabilityNoise for L1Sensitivity<TOC> {
+    fn noise(shift: Self::Distance, scale: Self::Distance, enforce_constant_time: bool) -> Fallible<Self::Distance> {
+        Self::Distance::sample_laplace(shift, scale, enforce_constant_time)
     }
 }
-impl<TOC: SampleGaussian> BaseStabilityNoise<TOC> for L2Sensitivity<TOC> {
-    fn noise(shift: TOC, scale: TOC, enforce_constant_time: bool) -> Fallible<TOC> {
-        TOC::sample_gaussian(shift, scale, enforce_constant_time)
+impl<TOC: SampleGaussian> BaseStabilityNoise for L2Sensitivity<TOC> {
+    fn noise(shift: Self::Distance, scale: Self::Distance, enforce_constant_time: bool) -> Fallible<Self::Distance> {
+        Self::Distance::sample_gaussian(shift, scale, enforce_constant_time)
     }
 }
 
-pub fn make_base_stability<MI, TIK, TIC, TOC>(
-    n: usize, scale: TOC, threshold: TOC
-) -> Fallible<Measurement<CountDomain<TIK, TIC>, CountDomain<TIK, TOC>, MI, SmoothedMaxDivergence<TOC>>>
-    where MI: BaseStabilityNoise<TOC> + SensitivityMetric<Distance=TOC>,
+pub fn make_base_stability<MI, TIK, TIC>(
+    n: usize, scale: MI::Distance, threshold: MI::Distance
+) -> Fallible<Measurement<CountDomain<TIK, TIC>, CountDomain<TIK, MI::Distance>, MI, SmoothedMaxDivergence<MI::Distance>>>
+    where MI: BaseStabilityNoise + SensitivityMetric,
           TIK: Eq + Hash + Clone,
           TIC: Integer + Clone + NumCast,
-          TOC: 'static + Float + Clone + PartialOrd + SampleLaplace + NumCast {
+          MI::Distance: 'static + Float + Clone + PartialOrd + NumCast {
     if scale.is_sign_negative() {
         return fallible!(MakeMeasurement, "scale must not be negative")
     }
     if threshold.is_sign_negative() {
         return fallible!(MakeMeasurement, "threshold must not be negative")
     }
-    let _n = c!(n; TOC)?;
-    let _2 = c!(2; TOC)?;
+    let _n = c!(n; MI::Distance)?;
+    let _2 = c!(2; MI::Distance)?;
 
     Ok(Measurement::new(
         SizedDomain::new(MapDomain { key_domain: AllDomain::new(), value_domain: AllDomain::new() }, n),
@@ -52,8 +52,8 @@ pub fn make_base_stability<MI, TIK, TIC, TOC>(
         Function::new_fallible(move |data: &HashMap<TIK, TIC>| {
             data.iter()
                 .map(|(k, c_in)| {
-                    // cast the value to TOC (output count)
-                    let c_out = TOC::from(c_in.clone()).ok_or_else(|| err!(FailedCast))?;
+                    // cast the value to MI::Distance (output count)
+                    let c_out = c!(c_in.clone(); MI::Distance)?;
                     // noise output count
                     Ok((k.clone(), MI::noise(c_out, scale, false)?))
                 })
@@ -64,7 +64,7 @@ pub fn make_base_stability<MI, TIK, TIC, TOC>(
         }),
         MI::default(),
         SmoothedMaxDivergence::default(),
-        PrivacyRelation::new_fallible(move |&d_in: &TOC, &(eps, del): &(TOC, TOC)|{
+        PrivacyRelation::new_fallible(move |&d_in: &MI::Distance, &(eps, del): &(MI::Distance, MI::Distance)|{
             // let _eps: f64 = NumCast::from(eps).unwrap_test();
             // let _del: f64 = NumCast::from(del).unwrap_test();
             // println!("eps, del: {:?}, {:?}", _eps, _del);
@@ -106,7 +106,7 @@ mod test_stability {
         let mut arg = HashMap::new();
         arg.insert(true, 6);
         arg.insert(false, 4);
-        let measurement = make_base_stability::<L2Sensitivity<f64>, bool, i8, f64>(10, 0.5, 1.)?;
+        let measurement = make_base_stability::<L2Sensitivity<f64>, bool, i8>(10, 0.5, 1.)?;
         let ret = measurement.function.eval(&arg)?;
         println!("stability eval: {:?}", ret);
